requirements.txt
# Core Scraping & API
telethon
requests
python-dotenv
loguru

# Data Handling
pandas
sqlalchemy
psycopg2-binary

# dbt (Data Transformation)
dbt-core

# Backend & API
fastapi
uvicorn

# Orchestration
dagster
dagster-postgres
dagit
dagster-webserver
# Computer Vision
opencv-python
torch
ultralytics


